Moved controller for detailed item and search scene to controllers package
