+++ ShoppIt/src/main/controllers/DetailedItemPopUpController.java
@@ -1,4 +1,4 @@
-package helpers;
+package controllers;
 
 import java.net.URL;
 import java.util.ResourceBundle;
