+++ ShoppIt/src/main/controllers/SearchPopUpController.java
@@ -1,4 +1,4 @@
-package helpers;
+package controllers;
 
 import java.io.IOException;
 import java.net.URL;
